Fixed potential future problem, update test classes

+++ Assignment 2/src/mru/tsc/model/BoardGame.java
@@ -16,7 +16,7 @@ public class BoardGame extends Toy {
 	 * @param numPlayers         the minimum and maximum number of players for the board game
 	 * @param designer           the names of who designed the board game
 	 */
-	public BoardGame(int SN, String name, String brand, double price, int availiableCount, String ageAppropriate, String numPlayers, String designer) {
+	public BoardGame(long SN, String name, String brand, double price, int availiableCount, String ageAppropriate, String numPlayers, String designer) {
 		super(SN,name,brand,price,availiableCount,ageAppropriate);
 		this.numPlayers = numPlayers;
 		this.designer = designer;
